Fixes bug 761609 - Added an option for date range type in TCBS, added unit tests.
